requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
orjson==3.9.10
packaging==23.2
pluggy==1.3.0
psycopg==3.1.14
psycopg-binary==3.1.14
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pytest==7.4.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
